+++ src/main/java/pl/futurecollars/invoicing/db/memory/InMemoryDatabase.java
@@ -9,18 +9,18 @@ import pl.futurecollars.invoicing.model.Invoice;
 
 public class InMemoryDatabase implements Database {
 
-    private final HashMap<Integer, Invoice> invoiceInMemoryDatabase = new HashMap<>();
-    private int index = 1;
+    private final HashMap<Long, Invoice> invoiceInMemoryDatabase = new HashMap<>();
+    private long index = 1;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         invoice.setId(index);
         invoiceInMemoryDatabase.put(index, invoice);
         return index++;
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         return Optional.ofNullable(invoiceInMemoryDatabase.get(id));
     }
 
@@ -30,13 +30,13 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         updatedInvoice.setId(id);
         return Optional.ofNullable(invoiceInMemoryDatabase.put(id, updatedInvoice));
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         return Optional.ofNullable(invoiceInMemoryDatabase.remove(id));
     }
 }
